requirements.txt
aiohttp==3.6.2
aiosqlite==0.15.0
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
# awsebcli==3.19.0
bcrypt==3.2.0
black==20.8b1
blessed==1.17.10
botocore==1.17.63
cached-property==1.5.2
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1.1
cycler==0.10.0
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus.git@84caae8038d0d3adc860957ccef05baeec2e2dd8
discord-py-slash-command==1.0.9.4
discord.py==1.6.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
flake8==3.8.3
future==0.16.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==4.3.21
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.0
mccabe==0.6.1
multidict==4.7.6
mypy-extensions==0.4.3
numpy==1.19.1
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.8.0
Pillow==8.1.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
semantic-version==2.5.0
six==1.15.0
termcolor==1.1.0
texttable==1.6.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.1.9
websocket-client==0.57.0
wrapt==1.12.1
yarl==1.5.1
